chore(upgrade-test): move js utils to lib/; copy into container
