+++ packages/deployment/upgrade-test/upgrade-test-scripts/agoric-upgrade-10/actions.test.js
@@ -6,11 +6,15 @@ import {
   raiseDebtCeiling,
   pushPrice,
 } from './actions.js';
-import { agd, agoric, agops } from '../cliHelper.js';
-import { GOV1ADDR, GOV2ADDR } from '../constants.js';
-import { getUser, newOfferId, waitForBlock } from '../commonUpgradeHelpers.js';
+import { agd, agoric, agops } from '../lib/cliHelper.js';
+import { GOV1ADDR, GOV2ADDR } from '../lib/constants.js';
+import {
+  getUser,
+  newOfferId,
+  waitForBlock,
+} from '../lib/commonUpgradeHelpers.js';
 import { submitDeliverInbound } from './upgradeHelpers.js';
-import { openVault, adjustVault, closeVault } from '../econHelpers.js';
+import { openVault, adjustVault, closeVault } from '../lib/econHelpers.js';
 
 const START_FREQUENCY = 600; // StartFrequency: 600s (auction runs every 10m)
 const CLOCK_STEP = 20; // ClockStep: 20s (ensures auction completes in time)
