+++ packages/deployment/upgrade-test/upgrade-test-scripts/agoric-upgrade-11/actions.test.js
@@ -1,8 +1,8 @@
 import test from 'ava';
 
-import { agoric, agops } from '../cliHelper.js';
-import { GOV1ADDR } from '../constants.js';
-import { openVault, adjustVault, closeVault } from '../econHelpers.js';
+import { agoric, agops } from '../lib/cliHelper.js';
+import { GOV1ADDR } from '../lib/constants.js';
+import { openVault, adjustVault, closeVault } from '../lib/econHelpers.js';
 
 test.serial('Open Vaults', async t => {
   const currentVaults = await agops.vaults('list', '--from', GOV1ADDR);
