+++ packages/deployment/upgrade-test/upgrade-test-scripts/lib/cliHelper.js
