+++ packages/deployment/upgrade-test/upgrade-test-scripts/lib/commonUpgradeHelpers.js
